fixed db connections (InsurancePolicy) in creating policy process and cast issue in Dummyscript

+++ src/main/java/de/thkoeln/inf/gpm/vgb/delegate/departmentclerk/DummyScriptTaskDelegate.java
@@ -34,7 +34,7 @@ public class DummyScriptTaskDelegate implements JavaDelegate {
         * Integer bmi
         **/
 
-        processVariables.put("alter", calcAge(geburtstag));
+        processVariables.put("alter", (long) calcAge(geburtstag));
         processVariables.put("bmi", bmi);
